fix: update css plugin
