+++ packages/core/src/provider/index.ts
@@ -5,7 +5,7 @@ export { initPlugins } from '@rsbuild/shared';
 export { initHooks, type Hooks } from './core/initHooks';
 export { initRsbuildConfig } from './core/initConfigs';
 export { getPluginAPI } from './core/initPlugins';
-export { applyBaseCSSRule, applyCSSModuleRule } from './plugins/css';
+export { applyBaseCSSRule } from './plugins/css';
 export type { InternalContext } from '../types';
 export { setHTMLPlugin, getHTMLPlugin } from './htmlPluginUtil';
 export { formatStats } from './shared';
