+++ packages/core/src/provider/plugins/css.ts
@@ -1,68 +1,79 @@
-import path from 'node:path';
 import {
-  logger,
-  getBrowserslistWithDefault,
-  isUseCssExtract,
   CSS_REGEX,
-  CSS_MODULES_REGEX,
-  getCssLoaderOptions,
-  getPostcssLoaderOptions,
-  getCssModuleLocalIdentName,
   resolvePackage,
+  isUseCssExtract,
+  getPostcssLoaderOptions,
   mergeChainedOptions,
+  getCssLoaderOptions,
+  getBrowserslistWithDefault,
+  getCssModuleLocalIdentName,
   getSharedPkgCompiledPath,
-  type BundlerChain,
   type RsbuildContext,
-  type RspackRule,
-  type RuleSetRule,
-  type ModifyBundlerChainUtils,
+  type RsbuildPlugin,
+  type NormalizedConfig,
+  type BundlerChainRule,
+  type ModifyChainUtils,
+  type CSSExtractOptions,
 } from '@rsbuild/shared';
-import type { RsbuildPlugin, NormalizedConfig } from '../../types';
-
-export const enableNativeCss = (config: NormalizedConfig) =>
-  !config.output.injectStyles;
+import { RspackCssExtractPlugin } from '@rspack/core';
 
 export async function applyBaseCSSRule({
   rule,
   config,
   context,
-  utils: { target, isProd, isServer, isWebWorker, CHAIN_ID },
+  utils: { target, isProd, isServer, CHAIN_ID, isWebWorker },
   importLoaders = 1,
 }: {
-  rule: ReturnType<BundlerChain['module']['rule']>;
+  rule: BundlerChainRule;
   config: NormalizedConfig;
   context: RsbuildContext;
-  utils: ModifyBundlerChainUtils;
+  utils: ModifyChainUtils;
   importLoaders?: number;
 }) {
-  // 1. Check user config
-  const enableCSSModuleTS = Boolean(config.output.enableCssModuleTSDeclaration);
-
   const browserslist = await getBrowserslistWithDefault(
     context.rootPath,
     config,
     target,
   );
 
-  // when disableExtractCSS, use css-loader + style-loader
-  if (!enableNativeCss(config)) {
-    const localIdentName = getCssModuleLocalIdentName(config, isProd);
+  // 1. Check user config
+  const enableExtractCSS = isUseCssExtract(config, target);
+  const enableCSSModuleTS = Boolean(config.output.enableCssModuleTSDeclaration);
 
-    const cssLoaderOptions = getCssLoaderOptions({
-      config,
-      importLoaders,
-      isServer,
-      isWebWorker,
-      localIdentName,
-    });
+  // 2. Prepare loader options
+  const localIdentName = getCssModuleLocalIdentName(config, isProd);
 
-    if (!isServer && !isWebWorker) {
+  const cssLoaderOptions = getCssLoaderOptions({
+    config,
+    importLoaders,
+    isServer,
+    isWebWorker,
+    localIdentName,
+  });
+
+  // 3. Create webpack rule
+  // Order: style-loader/mini-css-extract -> css-loader -> postcss-loader
+  if (!isServer && !isWebWorker) {
+    // use mini-css-extract-plugin loader
+    if (enableExtractCSS) {
+      const extraCSSOptions: Required<CSSExtractOptions> =
+        typeof config.tools.cssExtract === 'object'
+          ? config.tools.cssExtract
+          : {
+              loaderOptions: {},
+              pluginOptions: {},
+            };
+
+      rule
+        .use(CHAIN_ID.USE.MINI_CSS_EXTRACT)
+        .loader(RspackCssExtractPlugin.loader)
+        .options(extraCSSOptions.loaderOptions)
+        .end();
+    }
+    // use style-loader
+    else {
       const styleLoaderOptions = mergeChainedOptions({
-        defaults: {
-          // todo: hmr does not work while esModule is true
-          // @ts-expect-error
-          esModule: false,
-        },
+        defaults: {},
         options: config.tools.styleLoader,
       });
 
@@ -71,52 +82,35 @@ export async function applyBaseCSSRule({
         .loader(getSharedPkgCompiledPath('style-loader'))
         .options(styleLoaderOptions)
         .end();
-
-      // use css-modules-typescript-loader
-      if (enableCSSModuleTS && cssLoaderOptions.modules) {
-        rule
-          .use(CHAIN_ID.USE.CSS_MODULES_TS)
-          .loader(
-            resolvePackage(
-              '@rsbuild/shared/css-modules-typescript-loader',
-              __dirname,
-            ),
-          )
-          .options({
-            modules: cssLoaderOptions.modules,
-          })
-          .end();
-      }
-    } else {
-      rule
-        .use(CHAIN_ID.USE.IGNORE_CSS)
-        .loader(resolvePackage('@rsbuild/shared/ignore-css-loader', __dirname))
-        .end();
     }
 
-    rule
-      .use(CHAIN_ID.USE.CSS)
-      .loader(getSharedPkgCompiledPath('css-loader'))
-      .options(cssLoaderOptions)
-      .end();
-  } else {
-    // can not get experiment.css result, so we fake a css-modules-typescript-pre-loader
-    if (!isServer && !isWebWorker && enableCSSModuleTS) {
-      const { cssModules } = config.output;
+    // use css-modules-typescript-loader
+    if (enableCSSModuleTS && cssLoaderOptions.modules) {
       rule
         .use(CHAIN_ID.USE.CSS_MODULES_TS)
-        .loader(path.resolve(__dirname, '../css-modules-typescript-pre-loader'))
+        .loader(
+          resolvePackage(
+            '@rsbuild/shared/css-modules-typescript-loader',
+            __dirname,
+          ),
+        )
         .options({
-          modules: {
-            exportLocalsConvention: cssModules.exportLocalsConvention,
-            auto: cssModules.auto,
-          },
+          modules: cssLoaderOptions.modules,
         })
         .end();
     }
-
-    rule.type('css');
+  } else {
+    rule
+      .use(CHAIN_ID.USE.IGNORE_CSS)
+      .loader(resolvePackage('@rsbuild/shared/ignore-css-loader', __dirname))
+      .end();
   }
+
+  rule
+    .use(CHAIN_ID.USE.CSS)
+    .loader(getSharedPkgCompiledPath('css-loader'))
+    .options(cssLoaderOptions)
+    .end();
 
   if (!isServer && !isWebWorker) {
     const postcssLoaderOptions = await getPostcssLoaderOptions({
@@ -140,123 +134,28 @@ export async function applyBaseCSSRule({
   rule.resolve.preferRelative(true);
 }
 
-/**
- * Use type: "css/module" rule instead of css-loader modules.auto config
- *
- * applyCSSModuleRule in modifyRspackConfig, so that other plugins can easily adjust css rule in Chain.
- */
-export const applyCSSModuleRule = (
-  rules: RspackRule[] | undefined,
-  ruleTest: RegExp,
-  config: NormalizedConfig,
-) => {
-  if (!rules || !enableNativeCss(config)) {
-    return;
-  }
-
-  const ruleIndex = rules.findIndex(
-    (r) => r && r !== '...' && r.test === ruleTest,
-  );
-
-  if (ruleIndex === -1) {
-    return;
-  }
-
-  const cssModulesAuto = config.output.cssModules.auto;
-
-  if (!cssModulesAuto) {
-    return;
-  }
-
-  const rule = rules[ruleIndex] as RuleSetRule;
-
-  const { test, type, ...rest } = rule;
-
-  rules[ruleIndex] = {
-    test: ruleTest,
-    oneOf: [
-      {
-        ...rest,
-        test:
-          typeof cssModulesAuto !== 'boolean'
-            ? cssModulesAuto
-            : // auto: true
-              CSS_MODULES_REGEX,
-        type: 'css/module',
-      },
-      {
-        ...rest,
-        type: 'css',
-      },
-    ],
-  };
-};
-
 export const pluginCss = (): RsbuildPlugin => {
   return {
     name: 'rsbuild:css',
     setup(api) {
       api.modifyBundlerChain(async (chain, utils) => {
-        const config = api.getNormalizedConfig();
-
         const rule = chain.module.rule(utils.CHAIN_ID.RULE.CSS);
+        const config = api.getNormalizedConfig();
         rule.test(CSS_REGEX);
-
         await applyBaseCSSRule({
           rule,
           utils,
           config,
           context: api.context,
         });
-
-        const enableExtractCSS = isUseCssExtract(config, utils.target);
-
-        // TODO: there is no switch to turn off experiments.css sourcemap in rspack, so we manually remove css sourcemap in Rsbuild
-        if (!config.output.sourceMap.css && enableExtractCSS) {
-          const { RemoveCssSourcemapPlugin } = await import(
-            '../../rspack/RemoveCssSourcemapPlugin'
-          );
-          chain
-            .plugin('remove-css-sourcemap')
-            .use(RemoveCssSourcemapPlugin, []);
-        }
       });
-      api.modifyRspackConfig(
-        async (rspackConfig, { isProd, isServer, isWebWorker }) => {
-          const config = api.getNormalizedConfig();
 
-          if (!enableNativeCss(config)) {
-            rspackConfig.experiments ||= {};
-            rspackConfig.experiments.css = false;
-            return;
-          }
-
-          let localIdentName =
-            config.output.cssModules.localIdentName ||
-            // Using shorter classname in production to reduce bundle size
-            (isProd ? '[local]-[hash:6]' : '[path][name]__[local]-[hash:6]');
-
-          if (localIdentName.includes(':base64')) {
-            logger.warn(
-              `Custom hashDigest in output.cssModules.localIdentName is currently not supported when using Rspack, the 'base64' will be ignored.`,
-            );
-            localIdentName = localIdentName.replace(':base64', '');
-          }
-
-          // need use type: "css/module" rule instead of modules.auto config
-          rspackConfig.builtins ||= {};
-          rspackConfig.builtins.css ||= {};
-          rspackConfig.builtins.css.modules = {
-            localsConvention: config.output.cssModules.exportLocalsConvention,
-            localIdentName,
-            exportsOnly: isServer || isWebWorker,
-          };
-
-          const rules = rspackConfig.module?.rules;
-
-          applyCSSModuleRule(rules, CSS_REGEX, config);
-        },
-      );
+      api.modifyRspackConfig(async (rspackConfig) => {
+        rspackConfig.experiments ||= {};
+        rspackConfig.experiments.css = false;
+        rspackConfig.experiments.rspackFuture ||= {};
+        rspackConfig.experiments.rspackFuture.newTreeshaking = true;
+      });
     },
   };
 };
